+++ ts/tests/commands/precompile-test.js
@@ -18,7 +18,7 @@ describe('Acceptance: ts:precompile command', function() {
     return emberNew({ target: 'addon' }).then(() => ember(['generate', 'ember-cli-typescript']));
   });
 
-  it('generates .js and .d.ts files from the addon tree', function() {
+  it('generates .d.ts files from the addon tree', function() {
     fs.ensureDirSync('addon');
     fs.writeFileSync('addon/test-file.ts', `export const testString: string = 'hello';`);
 
@@ -27,34 +27,28 @@ describe('Acceptance: ts:precompile command', function() {
         let declaration = file('test-file.d.ts');
         expect(declaration).to.exist;
         expect(declaration.content.trim()).to.equal(`export declare const testString: string;`);
-
-        let transpiled = file('addon/test-file.js');
-        expect(transpiled).to.exist;
-        expect(transpiled.content.trim()).to.equal(`export const testString = 'hello';`);
       });
   });
 
-  it('generates .js files only from the app tree', function() {
+  it('generates nothing from the app tree', function() {
     fs.ensureDirSync('app');
     fs.writeFileSync('app/test-file.ts', `export const testString: string = 'hello';`);
 
     return ember(['ts:precompile']).then(() => {
       let declaration = file('test-file.d.ts');
       expect(declaration).not.to.exist;
-
-      let transpiled = file('app/test-file.js');
-      expect(transpiled).to.exist;
-      expect(transpiled.content.trim()).to.equal(`export const testString = 'hello';`);
     });
   });
 
   describe('module unification', function() {
-    it('generates .js and .d.ts files from the src tree', function() {
+    it('generates .d.ts files from the src tree', function() {
       fs.ensureDirSync('src');
       fs.writeFileSync('src/test-file.ts', `export const testString: string = 'hello';`);
 
+      let pkg = fs.readJsonSync('package.json');
       let tsconfig = fs.readJSONSync('tsconfig.json');
       tsconfig.include.push('src');
+      tsconfig.compilerOptions.paths[`${pkg.name}/src/*`] = ['src/*'];
       fs.writeJSONSync('tsconfig.json', tsconfig);
 
       return ember(['ts:precompile'])
@@ -62,10 +56,26 @@ describe('Acceptance: ts:precompile command', function() {
           let declaration = file('src/test-file.d.ts');
           expect(declaration).to.exist;
           expect(declaration.content.trim()).to.equal(`export declare const testString: string;`);
+        });
+    });
+  });
 
-          let transpiled = file('src/test-file.js');
-          expect(transpiled).to.exist;
-          expect(transpiled.content.trim()).to.equal(`export const testString = 'hello';`);
+  describe('remapped addon-test-support', function() {
+    it('generates .d.ts files in the mapped location', function() {
+      fs.ensureDirSync('addon-test-support');
+      fs.writeFileSync('addon-test-support/test-file.ts', `export const testString: string = 'hello';`);
+
+      let pkg = fs.readJsonSync('package.json');
+      let tsconfig = fs.readJSONSync('tsconfig.json');
+      tsconfig.include.push('src');
+      tsconfig.compilerOptions.paths[`${pkg.name}/*`] = ['addon-test-support/*'];
+      fs.writeJSONSync('tsconfig.json', tsconfig);
+
+      return ember(['ts:precompile'])
+        .then(() => {
+          let declaration = file('test-file.d.ts');
+          expect(declaration).to.exist;
+          expect(declaration.content.trim()).to.equal(`export declare const testString: string;`);
         });
     });
   });